+++ packages/melt/src/lib/components/index.ts
@@ -1,5 +1,8 @@
+export { default as Collapsible } from './Collapsible.svelte';
 export { default as Toggle } from "./Toggle.svelte";
 export { default as PinInput } from "./PinInput.svelte";
 export { default as Tabs } from "./Tabs.svelte";
 export { default as Slider } from "./Slider.svelte";
 export { default as Popover } from "./Popover.svelte";
+export { default as Progress } from "./Progress.svelte";
+export { default as RadioGroup } from "./RadioGroup.svelte";